requirements.txt
opencv-python>=4.10.0
pytesseract>=0.3.10
Pillow>=10.3.0
numpy>=1.26.0
python-dotenv>=1.0.1
Flask>=3.0.0
language-tool-python>=2.7.1
requests>=2.32.0
playwright>=1.45.0
sounddevice>=0.4.6
soundfile>=0.12.1
pyttsx3>=2.90
reportlab>=4.1.0
EbookLib>=0.18
python-docx>=1.1.2
pydub>=0.25.1
APScheduler>=3.10.4
openai>=1.40.0
supabase>=2.5.0
schedule>=1.2.2
